requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.26.0
sentence-transformers>=2.4.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
requests>=2.31.0
regex>=2023.8.8
scikit-learn>=1.3.0
google-generativeai>=0.7.0
argparse>=1.4.0 
